+++ android/app/src/main/java/com/trafi/example/OnboardingScreen.kt
@@ -15,7 +15,7 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.trafi.example.ui.DemoMaasTheme
-import com.trafi.ui.components.Button
+import com.trafi.ui.component.Button
 import com.trafi.ui.theme.MaasTheme
 
 @Composable
