+++ android/ui/src/main/java/com/trafi/ui/component/Button.kt
@@ -1,4 +1,4 @@
-package com.trafi.ui.components
+package com.trafi.ui.component
 
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.heightIn
