+++ android/ui/src/main/java/com/trafi/ui/component/OutlinedTextField.kt
@@ -1,4 +1,4 @@
-package com.trafi.ui.components
+package com.trafi.ui.component
 
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.clickable
